fix int parsing 64 (#204)
* fix int parsing 64

* use int64 in tests

* fix last test

+++ model/external_events.go
@@ -10,7 +10,7 @@ import (
 
 type ExternalEvent struct {
 	ID              int64     `db:"id"`
-	PageID          int64     `db:"page_id"`
+	PageID          int       `db:"page_id"`
 	Name            string    `db:"name"`
 	Description     string    `db:"description"`
 	StartTime       time.Time `db:"start_time"`
